requirements.txt
plotly==5.4.0
six==1.16.0
tenacity==8.0.1
openpyxl
matplotlib
streamlit_option_menu
